f4menu for lorerp by heavito
I made f4menu without vgui, but with hudpaint. I just didn’t know what it would be useful for, but looking at other addons I realized that it was very convenient. I’m merging what I managed to do, everything is buggy in this version, but you can roll back to the moment when the shop button did nothing and was displayed normally, and I went to make f4menu using vgui and part of the work in my menu!